upload: Move get_credentials() into UploadArea
